HTTP for Card and SSH ! (#58)
* add case for ssh

* fix doc

* fmt docs

* add card http + fmt doc

* add card http doc

* add ssh http

* add docs for ssh

* add gitignore

* fix credentials get

* add check validation credentials

* fix giti

* modify car number type + migrations

* some fix on check function

+++ credentials/migrations/000001_init.up.sql
@@ -15,7 +15,7 @@ CREATE TABLE IF NOT EXISTS card_credentials (
   owner_name VARCHAR(25) NOT NULL,
   CVC INTEGER NOT NULL,
   expiration_date VARCHAR(25) NOT NULL,
-  card_number INTEGER NOT NULL
+  card_number NUMERIC(16,0) NOT NULL
 ) INHERITS (credentials);
 
 CREATE TABLE IF NOT EXISTS password_credentials (
